emotion-detection-main/requirements.txt
absl-py==2.2.1
alembic==1.15.1
asgiref==3.7.2
astroid==3.3.9
astunparse==1.6.3
attrs==25.3.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
crispy-bootstrap4==2024.1
cycler==0.12.1
deepface==0.0.93
dill==0.3.9
distlib==0.3.8
Django==5.0.1
django-crispy-forms==2.1
filelock==3.13.1
fire==0.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
huggingface-hub==0.29.3
idna==3.6
isort==6.0.1
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.4.2
keras==3.9.1
kiwisolver==1.4.8
langdetect==1.0.9
libclang==18.1.1
lz4==4.4.3
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
mtcnn==1.0.0
mysql==0.0.3
mysqlclient==2.2.0
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pandas==2.2.3
pillow==10.2.0
platformdirs==4.2.0
protobuf==4.25.6
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.3.5
PyMySQL==1.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
razorpay==1.4.1
regex==2024.11.6
requests==2.31.0
retina-face==0.0.17
rich==13.9.4
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.39
sqlparse==0.4.4
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
textblob==0.19.0
tf_keras==2.19.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.1
vaderSentiment==3.3.2
virtualenv==20.25.0
Werkzeug==3.1.3
wrapt==1.17.2
groq==0.10.0
python-dotenv==1.0.0
